Indri core bug fix and namespace update.

+++ indri/branches/metzler/include/indri/QuerySpec.hpp
@@ -7,7 +7,7 @@
  * http://www.lemurproject.org/license.html
  *
  *==========================================================================
-*/
+ */
 
 
 #ifndef INDRI_QUERYSPEC_HPP
@@ -328,8 +328,8 @@ namespace indri {
         ExtentInside* other = dynamic_cast<ExtentInside*>(&o);
   
         return other &&
-               *_inner == *other->_inner &&
-               *_outer == *other->_outer;
+          *_inner == *other->_inner &&
+          *_outer == *other->_outer;
       }
       
       std::string typeName() const {
@@ -943,7 +943,7 @@ namespace indri {
           return false;
 
         return (*_filter) == (*other->getFilter()) &&
-               (*_required) == (*other->getRequired());
+          (*_required) == (*other->getRequired());
       }
 
       void pack( Packer& packer ) {
@@ -1016,7 +1016,7 @@ namespace indri {
           return false;
 
         return (*_filter) == (*other->getFilter()) &&
-               (*_disallowed) == (*other->getDisallowed());
+          (*_disallowed) == (*other->getDisallowed());
       }
 
       void pack( Packer& packer ) {
@@ -1080,8 +1080,8 @@ namespace indri {
         FieldLessNode* other = dynamic_cast<FieldLessNode*>(&node);
 
         return other &&
-               other->getConstant() == _constant &&
-               *other->getField() == *_field;
+          other->getConstant() == _constant &&
+          *other->getField() == *_field;
       }
 
       Node* copy( Copier& copier ) {
@@ -1143,8 +1143,8 @@ namespace indri {
         FieldGreaterNode* other = dynamic_cast<FieldGreaterNode*>(&node);
 
         return other &&
-               other->getConstant() == _constant &&
-               *other->getField() == *_field;
+          other->getConstant() == _constant &&
+          *other->getField() == *_field;
       }
 
       Node* copy( Copier& copier ) {
@@ -1213,9 +1213,9 @@ namespace indri {
         FieldBetweenNode* other = dynamic_cast<FieldBetweenNode*>(&node);
 
         return other &&
-               other->getLow() == _low &&
-               other->getHigh() == _high &&
-               *other->getField() == *_field;
+          other->getLow() == _low &&
+          other->getHigh() == _high &&
+          *other->getField() == *_field;
       }
 
       Node* copy( Copier& copier ) {
@@ -1276,11 +1276,11 @@ namespace indri {
       }
       
       bool operator== ( Node& node ) {
-	FieldEqualsNode* other = dynamic_cast<FieldEqualsNode*>(&node);
-	
-	return other &&
-	  other->getConstant() == _constant &&
-	  *other->getField() == *_field;
+        FieldEqualsNode* other = dynamic_cast<FieldEqualsNode*>(&node);
+
+        return other &&
+          other->getConstant() == _constant &&
+          *other->getField() == *_field;
       }
       
       Node* copy( Copier& copier ) {
@@ -1692,11 +1692,11 @@ namespace indri {
         for( std::map<int,tuple_type>::iterator iter;
              iter != _table.end();
              iter++ )
-        {
-          beginList.push_back( (*iter).second.begin );
-          endList.push_back( (*iter).second.end );
-          scoreList.push_back( (*iter).second.score );
-        }
+          {
+            beginList.push_back( (*iter).second.begin );
+            endList.push_back( (*iter).second.end );
+            scoreList.push_back( (*iter).second.score );
+          }
 
         packer.put( "begin", beginList );
         packer.put( "end", endList );
@@ -2391,19 +2391,19 @@ namespace indri {
       RawExtentNode* _context;
       bool _hasCounts;
       bool _hasContextSize;
-      UINT64 _occurrences;
-      UINT64 _contextSize;
+      double _occurrences;
+      double _contextSize;
 
       UINT64 _docOccurrences;
       int _docCount;
     public:
       ContextCounterNode( RawExtentNode* raw, RawExtentNode* context ) :
-         _hasCounts(false),
-         _hasContextSize(false),
-         _occurrences(0),
-         _contextSize(0),
-         _docOccurrences(0),
-         _docCount(0)
+        _hasCounts(false),
+        _hasContextSize(false),
+        _occurrences(0),
+        _contextSize(0),
+	_docOccurrences(0),
+	_docCount(0)
       {
         _raw = raw;
         _context = context;
@@ -2412,8 +2412,8 @@ namespace indri {
       ContextCounterNode( Unpacker& unpacker ) {
         _raw = unpacker.getRawExtentNode( "raw" );
         _context = unpacker.getRawExtentNode( "context" );
-        _occurrences = unpacker.getInteger( "occurrences" );
-        _contextSize = unpacker.getInteger( "contextSize" );
+        _occurrences = unpacker.getDouble( "occurrences" );
+        _contextSize = unpacker.getDouble( "contextSize" );
 
         _hasCounts = unpacker.getBoolean( "hasCounts" );
         _hasContextSize = unpacker.getBoolean( "hasContextSize" );
@@ -2508,13 +2508,13 @@ namespace indri {
         return _contextSize;
       }
 
-      void setContextSize( UINT64 contextSize ) {
+      void setContextSize( double contextSize ) {
         _contextSize = contextSize;
         _hasContextSize = true;
       }
 
-      void setCounts( UINT64 occurrences,
-                      UINT64 contextSize ) {
+      void setCounts( double occurrences,
+                      double contextSize ) {
         _hasCounts = true;
         _occurrences = occurrences;
         setContextSize( contextSize );
@@ -2529,6 +2529,7 @@ namespace indri {
 
       bool _hasCounts;
       bool _hasContextSize;
+
       double _occurrences;
       double _contextSize;
       
@@ -2537,19 +2538,19 @@ namespace indri {
       
     public:
       ContextSimpleCounterNode( const std::vector<std::string>& terms, const std::string& field, const std::string& context ) :
-         _hasCounts(false),
-         _hasContextSize(false),
-         _occurrences(0),
-         _contextSize(0),
-         _terms(terms),
-         _field(field),
-         _context(context)
+        _hasCounts(false),
+        _hasContextSize(false),
+        _occurrences(0),
+        _contextSize(0),
+        _terms(terms),
+        _field(field),
+        _context(context)
       {
       }
 
       ContextSimpleCounterNode( Unpacker& unpacker ) {
-        _occurrences = unpacker.getInteger( "occurrences" );
-        _contextSize = unpacker.getInteger( "contextSize" );
+        _occurrences = unpacker.getDouble( "occurrences" );
+        _contextSize = unpacker.getDouble( "contextSize" );
 
         _terms = unpacker.getStringVector( "terms" );
         _field = unpacker.getString( "field" );
@@ -2628,20 +2629,20 @@ namespace indri {
       }
 
       UINT64 getDocOccurrences() const {
-		return _docOccurrences;
+	return _docOccurrences;
       }
       
       int getDocCount() const {
-		return _docCount;
+	return _docCount;
       }
-
-      void setContextSize( UINT64 contextSize ) {
+      
+      void setContextSize( double contextSize ) {
         _contextSize = contextSize;
         _hasContextSize = true;
       }
 
-      void setCounts( UINT64 occurrences,
-                      UINT64 contextSize,
+      void setCounts( double occurrences,
+                      double contextSize,
                       UINT64 docOccurrences ) {
         _hasCounts = true;
         _occurrences = occurrences;
@@ -2700,55 +2701,52 @@ namespace indri {
     };
 
     class AnnotatorNode : public AccumulatorNode {
-      private:
-        ScoredExtentNode* _scoredNode;
+    private:
+      ScoredExtentNode* _scoredNode;
 
-      public:
-        AnnotatorNode( ScoredExtentNode* scoredNode ) :
-          _scoredNode(scoredNode)
-        {
-        }
+    public:
+      AnnotatorNode( ScoredExtentNode* scoredNode ) :
+        _scoredNode(scoredNode)
+      {
+      }
 
-        AnnotatorNode( Unpacker& unpacker ) {
-          _scoredNode = unpacker.getScoredExtentNode( "scoredNode" );
-        }
+      AnnotatorNode( Unpacker& unpacker ) {
+        _scoredNode = unpacker.getScoredExtentNode( "scoredNode" );
+      }
 
-        std::string typeName() const {
-          return "AnnotatorNode";
-        }
+      std::string typeName() const {
+        return "AnnotatorNode";
+      }
 
-        std::string queryText() const {
-          return _scoredNode->queryText();
-        }
+      std::string queryText() const {
+        return _scoredNode->queryText();
+      }
 
-        ScoredExtentNode* getChild() {
-          return _scoredNode;
-        }
+      ScoredExtentNode* getChild() {
+        return _scoredNode;
+      }
 
-        void pack( Packer& packer ) {
-          packer.before(this);
-          packer.put( "scoredNode", _scoredNode );
-          packer.after(this);
-        }
+      void pack( Packer& packer ) {
+        packer.before(this);
+        packer.put( "scoredNode", _scoredNode );
+        packer.after(this);
+      }
 
-        void walk( Walker& walker ) {
-          walker.before(this);
-          _scoredNode->walk(walker);
-          walker.after(this);
-        }
+      void walk( Walker& walker ) {
+        walker.before(this);
+        _scoredNode->walk(walker);
+        walker.after(this);
+      }
 
-        Node* copy( Copier& copier ) {
-          copier.before(this);
-          ScoredExtentNode* duplicateChild = dynamic_cast<ScoredExtentNode*>(_scoredNode->copy(copier));
-          AnnotatorNode* duplicate = new AnnotatorNode(duplicateChild);
-          duplicate->setNodeName( nodeName() );
-          return copier.after(this, duplicate);
-        }
+      Node* copy( Copier& copier ) {
+        copier.before(this);
+        ScoredExtentNode* duplicateChild = dynamic_cast<ScoredExtentNode*>(_scoredNode->copy(copier));
+        AnnotatorNode* duplicate = new AnnotatorNode(duplicateChild);
+        duplicate->setNodeName( nodeName() );
+        return copier.after(this, duplicate);
+      }
     };
   }
 }
 
 #endif // INDRI_QUERYSPEC_HPP
-
-
-
